requirements.txt
apache-airflow==2.7.3
apache-airflow-providers-docker
apache-airflow-providers-postgres
pyspark==3.5.0
boto3==1.34.7
minio==7.2.3
pandas==2.1.4
pyarrow==14.0.2
python-dotenv==1.0.0
requests==2.31.0
dbt-core==1.7.4
dbt-postgres==1.7.4
